requirement.txt
prometheus_client==0.20.0
psycopg2-binary==2.9.9
requests==2.32.3
schedule==1.2.2
python-dotenv==1.0.1